+++ src/lib/ltree/ltree-node.svelte.ts
@@ -62,7 +62,7 @@ export function createLTreeNode<T>(data?: Partial<LTreeNode<T>>): LTreeNode<T> {
 		isInsertAllowed: true,
 		isNestAllowed: true,
 		isCheckboxVisible: false,
-		visualState: VisualState.indeterminate,
+		visualState: VisualState.notSelected,
 		isExpanded: false,
 		isSelected: false,